Initial ARM64 build support
International support causes a bunch of problems because AL.exe doesn't support AMR64 as a valid platform. The work around is to disable the built in assembly linking and replicate the action using a cusotm action. Another option is to compile for x64 and then copy the needed assemblies to the correct directories as mentioned in the following stackoverflow answer:
https://stackoverflow.com/questions/77414978/c-sharp-wpf-implements-multi-language-internationalization-support-and-the-corr